Use eth0 to serve for DHCP clients instead of eth1.
This allows DreamPlugs to work smoothly.  eth0 is used as a DHCP
server for incoming (downstream) clients and eth1 is a DHCP client for
an outgoing (upstream) server.  Now, the dnsmasq server interface is
correctly eth0: otherwise, we'd try to send outgoing packets back to
the client, who'd reject it.

: (computer) -> eth0 -> (dnsmasq) -> eth1 -> (Internet)

eth1 is the outer Ethernet port on a DreamPlug.