Reverse roles of eth0 and eth1 ports.
Now:

- eth0 :: DHCP client (connects to router).
- eth1 :: DHCP server (connects to end-device).

This allows all targeted hardware to connect to the Internet without
issue, whether or not they have an eth1 port.

All devices will be DHCP clients on eth0 so all devices can connect
upstream, to the Internet and other FreedomBoxes.  Only devices with
an eth1 port will be able to act as routers.