Renamed NOSOURCE variable to SOURCE.
I don't like anti-options :)

This should be a transparent change because of
setup.d/99_provide-source::4, which handles both the old and new
variable names:

    -if [ 'true' = "$NOSOURCE" ] ; then
    +if [ 'false' = "$SOURCE" ] || [ 'true' = '$NOSOURCE' ] ; then

This way, we do the correct thing even if the user is using an
outdated version of freedom-maker that hasn't renamed NOSOURCE yet.
This can be cleaned up later, but after we're sure people have
transitioned to an updated FM version.  By my count, they should have
till the end of the year.  Realistically, everybody who cares should
be upgraded within a few weeks to a month at most.