Move Apache configuration to a file
If Apache configuration is output as a file, it makes upgrades to that
file difficult when newer version of freedombox-setup is avialable.
Moving to a file managed by Debian infrastucture solves that issue.

Also place the file as a available configuration rather than available
site as a new site is not created.

Signed-off-by: James Valleroy <jvalleroy@mailbox.org>