Don't delete SSH keys
When users install FreedomBox on a Debian system, it should not wipe out
their existing SSH server keys.  This should only happen in an pre-build
image downloadable by many people so as to not have multiple servers
having the same set of keys.

This patch complements a change to freedom-maker that makes sure the SSH
server keys are not present on a built image.

Signed-off-by: James Valleroy <jvalleroy@mailbox.org>